+++ include/generated_info.hpp
@@ -1,6 +1,6 @@
 // ---- Generated by scripts/gen_info_header.py ----
 #pragma once
-#include "data.hpp"
+#include "enums.hpp"
 
 #include <array>
 #include <set>
