+++ scripts/gen_info_header.py
@@ -95,7 +95,7 @@ def main(out_folder: Path = Path.cwd()):
     """Generate C++ Game Info Header based on values from PySC2"""
     content = r"""// ---- Generated by scripts/gen_info_header.py ----
 #pragma once
-#include "data.hpp"
+#include "enums.hpp"
 
 #include <array>
 #include <set>
