+++ test/test_readwrite.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 #include <spdlog/spdlog.h>
 
-#include "data.hpp"
+#include "replay_structures.hpp"
 #include "serialize.hpp"
 
 #include <filesystem>
@@ -21,14 +21,14 @@ class ReplayDataTest : public testing::Test
         for (int i = 0; i < 3; ++i) {
             cvt::Action action = { .unit_ids = { 1, 2, 3 },
                 .ability_id = 6,
-                .target_type = cvt::Action::Target_Type::OtherUnit,
+                .target_type = cvt::Action::TargetType::OtherUnit,
                 .target = cvt::Action::Target(static_cast<cvt::UID>(3)) };
             replay_.stepData[0].actions.emplace_back(std::move(action));
         }
         for (int i = 0; i < 3; ++i) {
             cvt::Action action = { .unit_ids = { 1, static_cast<cvt::UID>(i) },
                 .ability_id = 1,
-                .target_type = cvt::Action::Target_Type::Position,
+                .target_type = cvt::Action::TargetType::Position,
                 .target = cvt::Action::Target(cvt::Point2d(i, 2)) };
             replay_.stepData[0].actions.emplace_back(std::move(action));
         }
